Fixed authorization(with frontend needs). Added admin's verifying.

+++ backend/api.py
@@ -1,7 +1,6 @@
 from fastapi import FastAPI, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from jose import JWTError, jwt
 from passlib.context import CryptContext
 from typing import Annotated
@@ -43,8 +42,6 @@ SECRET_KEY = "0b2ab43e07f7d350574ed3779b48409a442ab98d14f51511dc7a4b61c359b033"
 ALGORITHM = "HS256"
 ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="autorization/token")
-
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 
@@ -59,13 +56,15 @@ def get_password_hash(password):
 
 def authenticate_user(login: str, password: str, db: Session = Depends(get_db)):
     user = crud.get_operator_by_login(db=db, operator_login=login)
+    role = "operator" if user is not None else None
     if user is None:
         user = crud.get_admin_by_login(db=db, admin_login=login)
+        role = "admin"
     if user is None:
-        return False
+        return (False, False)
     if not verify_password(password, user.password):
-        return False
-    return user
+        return (False, False)
+    return (user, role)
 
 
 def create_access_token(data: dict, expires_delta: timedelta | None = None):
@@ -79,12 +78,12 @@ def create_access_token(data: dict, expires_delta: timedelta | None = None):
     return encoded_jwt
 
 
-def get_current_user(token: Annotated[str, Depends(oauth2_scheme)], db: Session = Depends(get_db)):
+def get_current_user(token: schemas.Token, db: Session = Depends(get_db)):
     credentials_exception = HTTPException(
         status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
+        detail="Could not validate credentials"
     )
+    token = token.access_token
     try:
         payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
         username: str = payload.get("sub")
@@ -109,39 +108,39 @@ def navigation_route():
 
 # Авторизация пользователя.
 @app.post("/autorization/token", response_model=schemas.Token, tags=["autorization"])
-def login_for_access_token(form_data: Annotated[OAuth2PasswordRequestForm, Depends()], db: Session = Depends(get_db)):
-    user = authenticate_user(form_data.username, form_data.password, db=db)
+def login_for_access_token(form_data: schemas.LoginFormData, db: Session = Depends(get_db)):
+    user, role = authenticate_user(login=form_data.username, password=form_data.password, db=db)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Bearer"},
+            detail="Incorrect username or password"
         )
     access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
     access_token = create_access_token(
         data={"sub": user.login}, expires_delta=access_token_expires
     )
-    return {"access_token": access_token, "token_type": "bearer"}
-
-
-@app.get("/autorization/active_admin", response_model=schemas.Admin, tags=["autorization"])
-def admin(active_admin: Annotated[schemas.Admin, Depends(get_current_user)]):
-    return active_admin
+    return {"access_token": access_token, "role": role}
 
 
 # Operator API.
 @app.get("/operators/get_all", response_model=list[schemas.OperatorPrivate], tags=["operators"])
-def get_all_operators(active_admin: Annotated[schemas.Admin, Depends(get_current_user)], db: Session = Depends(get_db)):
+def get_all_operators(db: Session = Depends(get_db)):
     return crud.get_all_operators(db=db)
 
 
 @app.post("/operators/create", response_model=schemas.OperatorPrivate, tags=["operators"])
-def create_operator(operator: schemas.OperatorCreate, db: Session = Depends(get_db)):
+def create_operator(operator: schemas.OperatorCreate, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     return crud.create_operator(db=db, operator=operator)
 
 
 @app.post("/operators/update", response_model=schemas.OperatorPrivate, tags=["operators"])
-def update_operator(operator: schemas.OperatorUpdate, db: Session = Depends(get_db)):
+def update_operator(operator: schemas.OperatorUpdate, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     operator = crud.update_operator(db=db, operator_update=operator)
     if operator is None:
         raise HTTPException(status_code=404, detail="It is nothing to update!")
@@ -149,7 +148,10 @@ def update_operator(operator: schemas.OperatorUpdate, db: Session = Depends(get_
 
 
 @app.post("/operators/update_login", response_model=schemas.OperatorPrivate, tags=["operators"])
-def update_operator_login(operator: schemas.OperatorUpdateLogin, db: Session = Depends(get_db)):
+def update_operator_login(operator: schemas.OperatorUpdateLogin, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     operator = crud.update_operator_login(db=db, operator_update=operator)
     if operator is None:
         raise HTTPException(status_code=404, detail="It is nothing to update!")
@@ -157,7 +159,10 @@ def update_operator_login(operator: schemas.OperatorUpdateLogin, db: Session = D
 
 
 @app.post("/operators/update_password", response_model=schemas.OperatorPrivate, tags=["operators"])
-def update_operator_password(operator: schemas.OperatorUpdatePassword, db: Session = Depends(get_db)):
+def update_operator_password(operator: schemas.OperatorUpdatePassword, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     operator = crud.update_operator_password(db=db, operator_update=operator)
     if operator is None:
         raise HTTPException(status_code=404, detail="It is nothing to update!")
@@ -165,7 +170,10 @@ def update_operator_password(operator: schemas.OperatorUpdatePassword, db: Sessi
 
 
 @app.post("/operators/delete", tags=["operators"])
-def delete_operator(operator_delete: schemas.OperatorDelete, db: Session = Depends(get_db)):
+def delete_operator(operator_delete: schemas.OperatorDelete, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     res = crud.delete_operator(db=db, operator_delete=operator_delete)
     if res is None:
         raise HTTPException(status_code=404, detail="It is nothing to delete!")
@@ -173,7 +181,8 @@ def delete_operator(operator_delete: schemas.OperatorDelete, db: Session = Depen
 
 
 @app.post("/operators/delete_list", tags=["operators"])
-def delete_operator_list(operator_delete_list: list[schemas.OperatorDelete], db: Session = Depends(get_db)):
+def delete_operator_list(operator_delete_list: list[schemas.OperatorDelete], current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
     return crud.delete_operator_list(db=db, operator_delete_list=operator_delete_list)
 
 
@@ -184,12 +193,18 @@ def get_all_hotels(db: Session = Depends(get_db)):
 
 
 @app.post("/hotels/create", response_model=schemas.Hotel, tags=["hotels"])
-def create_hotel(hotel: schemas.HotelCreate, db: Session = Depends(get_db)):
+def create_hotel(hotel: schemas.HotelCreate, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     return crud.create_hotel(db=db, hotel=hotel)
 
 
 @app.post("/hotels/update", response_model=schemas.Hotel, tags=["hotels"])
-def update_hotel(hotel: schemas.HotelUpdate, db: Session = Depends(get_db)):
+def update_hotel(hotel: schemas.HotelUpdate, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     hotel = crud.update_hotel(db=db, hotel_update=hotel)
     if hotel is None:
         raise HTTPException(status_code=404, detail="It is nothing to update!")
@@ -197,14 +212,20 @@ def update_hotel(hotel: schemas.HotelUpdate, db: Session = Depends(get_db)):
 
 
 @app.post("/hotels/delete", tags=["hotels"])
-def delete_hotel(hotel_delete: schemas.HotelDelete, db: Session = Depends(get_db)):
+def delete_hotel(hotel_delete: schemas.HotelDelete, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     res = crud.delete_hotel(db=db, hotel_delete=hotel_delete)
     if res is None:
         raise HTTPException(status_code=404, detail="It is nothing to delete!")
 
 
 @app.post("/hotels/delete_list", tags=["hotels"])
-def delete_hotel_list(hotel_delete_list: list[schemas.HotelDelete], db: Session = Depends(get_db)):
+def delete_hotel_list(hotel_delete_list: list[schemas.HotelDelete], current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     crud.delete_hotel_list(db=db, hotel_delete_list=hotel_delete_list)
 
 
@@ -215,13 +236,18 @@ def get_all_hotel_rooms(db: Session = Depends(get_db)):
 
 
 @app.post("/hotel_rooms/create", response_model=schemas.HotelRoom, tags=["rooms"])
-def create_hotel_room(hotel_room: schemas.HotelRoomCreate, db: Session = Depends(get_db)):
-    print(hotel_room)
+def create_hotel_room(hotel_room: schemas.HotelRoomCreate, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     return crud.create_hotel_room(db=db, hotel_room=hotel_room)
 
 
 @app.post("/hotel_rooms/update", response_model=schemas.HotelRoom, tags=["rooms"])
-def update_hotel_room(hotel_room: schemas.HotelRoomUpdate, db: Session = Depends(get_db)):
+def update_hotel_room(hotel_room: schemas.HotelRoomUpdate, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     hotel_room = crud.update_hotel_room(db=db, hotel_room_update=hotel_room)
     if hotel_room is None:
         raise HTTPException(status_code=404, detail="It is nothing to update!")
@@ -229,14 +255,20 @@ def update_hotel_room(hotel_room: schemas.HotelRoomUpdate, db: Session = Depends
 
 
 @app.post("/hotel_rooms/delete", tags=["rooms"])
-def delete_hotel_room(hotel_room_delete: schemas.HotelRoomDelete, db: Session = Depends(get_db)):
+def delete_hotel_room(hotel_room_delete: schemas.HotelRoomDelete, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     res = crud.delete_hotel_room(db=db, hotel_room_delete=hotel_room_delete)
     if res is None:
         raise HTTPException(status_code=404, detail="It is nothing to delete!")
 
 
 @app.post("/hotel_rooms/delete_list", tags=["rooms"])
-def delete_hotel_room_list(hotel_room_delete_list: list[schemas.HotelRoomDelete], db: Session = Depends(get_db)):
+def delete_hotel_room_list(hotel_room_delete_list: list[schemas.HotelRoomDelete], current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     return crud.delete_hotel_room_list(db=db, hotel_room_delete_list=hotel_room_delete_list)
 
 
@@ -263,12 +295,18 @@ def update_tour(tour: schemas.TourUpdate, db: Session = Depends(get_db)):
 
 
 @app.post("/tours/delete", tags=["tours"])
-def delete_tour(tour_delete: schemas.TourDelete, db: Session = Depends(get_db)):
+def delete_tour(tour_delete: schemas.TourDelete, current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     res = crud.delete_tour(db=db, tour_delete=tour_delete)
     if res is None:
         raise HTTPException(status_code=404, detail="It is nothing to delete!")
 
 
 @app.post("/tours/delete_list", tags=["tours"])
-def delete_tour_list(tour_delete_list: list[schemas.TourDelete], db: Session = Depends(get_db)):
+def delete_tour_list(tour_delete_list: list[schemas.TourDelete], current_user: Annotated[schemas.Token, Depends(get_current_user)],
+        db: Session = Depends(get_db)):
+    if not("Admin" in str(current_user)):
+        raise HTTPException(status_code=405)
     return crud.delete_tour_list(db=db, tour_delete_list=tour_delete_list)
